Pequeños cambios en css y control de eliminación

+++ src/app/page/islas/ruta-list/ruta-detail/ruta-detail.page.ts
@@ -3,6 +3,7 @@ import {ActivatedRoute, Router} from '@angular/router';
 import {RutaService} from '../../../../services/ruta.service';
 import {Ruta} from '../../../../services/ruta.model';
 import {ToastController} from '@ionic/angular';
+import {AuthenticateService} from '../../../../services/authenntication.service';
 
 @Component({
   selector: 'app-ruta-detail',
@@ -21,7 +22,8 @@ export class RutaDetailPage implements OnInit {
   constructor(private activatedRoute: ActivatedRoute,
               private rutaService: RutaService,
               private router: Router,
-              private toastController: ToastController) { }
+              private toastController: ToastController,
+              private auth: AuthenticateService) { }
 
   ngOnInit() {
     const id = this.activatedRoute.snapshot.paramMap.get('id');
@@ -33,14 +35,20 @@ export class RutaDetailPage implements OnInit {
   }
 
   deleteRuta(){
-    this.rutaService.deleteruta(this.ruta.id, this.ruta.urlFoto).then(() => {
-      this.router.navigateByUrl('/home/' + this.ruta.isla);
-      this.showToast('Ruta eliminada');
-    }, err => {
-      this.showToast('No se pudo eliminar la ruta');
-      console.log(err);
-    });
+    if (this.auth.isLogged){
+      this.rutaService.deleteruta(this.ruta.id, this.ruta.urlFoto).then(() => {
+        this.router.navigateByUrl('/home/' + this.ruta.isla);
+        this.showToast('Ruta eliminada');
+      }, err => {
+        const msg = 'Error: '  + err;
+        this.showToast(msg);
+        console.log(err);
+      });
+    } else {
+      this.showToast('El usuario debe estar registrado');
+    }
   }
+
   showToast(msg) {
     this.toastController.create({
       message: msg,
